==> HiFi_assemblies/C_amplexicaulis_S_pinnata_comparisons/scripts/05_WGD_Retention_Loss.Rmd <==
---
title: "WGD Retention and Loss"
output: html_document
date: "2024-11-19"
---

The goal is to identify those genes that have been retained as duplicates vs those that have had a loss and are now singletons, and then to do GO and KEGG enrichment on them.  I will use the GENESPACE analysis as input.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(topGO)
library(clusterProfiler)
conflicted::conflict_prefer("select", "dplyr")
conflicted::conflict_prefer("filter", "dplyr")
```
Load the gene info.  
```{r}
pangenes <- read_tsv("../output/genespace_no_pinnata/pangenes/S_diversifolius_pangenes.txt.gz")
pangenes
```

```{r}
pangenes %>% arrange(og)
```

```{r}
og.copyn <- pangenes %>% filter(flag == "PASS") %>%
  distinct(og, id, genome) %>% count(og, genome)
og.copyn
```


```{r}
og.copyn %>%
  filter(n < 6) %>%
  ggplot(aes(x=genome, fill=as.character(n))) +
  geom_bar()
```

```{r}
og.category <- og.copyn %>% 
  pivot_wider(names_from = genome, values_from = n, values_fill = 0) %>%
  mutate(retained.both = S_diversifolius==2 & C_amplexicaulis==2 & ((E_salsugineum==1 & S_parvula==1) | (E_salsugineum==1 & S_parvula==0) | (E_salsugineum==0 & S_parvula==1) ) ,
         retained.S_div = S_diversifolius==2 & C_amplexicaulis==1 & ((E_salsugineum==1 & S_parvula==1) | (E_salsugineum==1 & S_parvula==0) | (E_salsugineum==0 & S_parvula==1) ),
         retained.C_amp = S_diversifolius==1  & C_amplexicaulis==2 & ((E_salsugineum==1 & S_parvula==1) | (E_salsugineum==1 & S_parvula==0) | (E_salsugineum==0 & S_parvula==1) ),
         retained.neither = S_diversifolius==1 & C_amplexicaulis==1 &((E_salsugineum==1 & S_parvula==1) | (E_salsugineum==1 & S_parvula==0) | (E_salsugineum==0 & S_parvula==1) )) %>%
 # select(og, matches("retained")) %>%
  rowwise() %>%
  mutate(other = !any(c_across(matches("retained")))) %>%
  ungroup()

og.category
```

Add S_div gene id back
```{r}
og.category <- 
  
  pangenes %>% 
  filter(genome=="S_diversifolius") %>%
  select(og, id) %>%
  distinct() %>%
  # pull(id) %>% duplicated() %>% sum() # 0
  group_by(og) %>%
  mutate(index=row_number()) %>%
  ungroup() %>%
  #pull(id) %>% duplicated() %>% sum() # 0
  pivot_wider(names_from = index, names_prefix = "gene_", values_from = id) %>%
  select(og, gene_1, gene_2) %>%
  # pull(gene_2) %>% na.omit() %>% duplicated %>% sum # 0
  right_join(og.category, by = "og") %>%
  filter(!(is.na(gene_1) & is.na(gene_2))) %>%
  arrange(og)

og.category
```

```{r}
og.category %>% filter(other)
```


```{r}
og.category %>%
  dplyr::select(og, matches("retained|other")) %>%
  pivot_longer(-og, names_to = "category", values_to = "status") %>%
  filter(status) %>%
  count(category)
```

```{r}
og.category %>% write_csv("../output/genespace_no_pinnata/WGD_retention_loss_categories.csv")
```

## GO analysis

```{r}
load("../../HiFiAsm_S_div_gene2GO.Rdata")
go.table <- read_csv("../../HiFiasm_GO.csv.gz")
```

### how concordant are the GO terms across paralogs?

```{r}
og.go.compare <- og.category %>% 
  filter(S_diversifolius==2) %>%
  select(og, gene_1, gene_2) %>%
  pivot_longer(-og, names_to = "name", values_to = "ID") %>%
  left_join(go.table) %>%
  group_by(og, name) %>%
  nest(GO=GO) %>%
  pivot_wider(names_from = name, values_from = c(ID, GO)) %>%
  rowwise() %>%
  mutate(setequal = dplyr::setequal(GO_gene_1, GO_gene_2),
         nGO1 = nrow(GO_gene_1),
         nGO2 = nrow(GO_gene_2),
         total.diff = nrow(symdiff(GO_gene_1, GO_gene_2)),
         pct.diff = 100 * total.diff / (nGO1 + nGO2)
         ) %>%
  ungroup()

og.go.compare

sum(og.go.compare$setequal) / nrow(og.go.compare)

og.go.compare %>% filter(!setequal) %>% pull(pct.diff) %>% mean()

```
Not so concordant...

For each syntenic paralog in S.div, let's combine the GO terms

```{r}
og.union <- og.category %>% 
  filter(S_diversifolius==2) %>%
  select(og, gene_1, gene_2) %>%
  pivot_longer(-og, names_to = "name", values_to = "ID") %>%
  left_join(go.table) %>%
  group_by(og, name) %>%
  nest(GO=GO) %>%
  pivot_wider(names_from = name, values_from = c(ID, GO)) %>%
  rowwise() %>%
  mutate(GOunion = list(na.omit(dplyr::union(GO_gene_1, GO_gene_2)))) %>%
  ungroup()

# check it
og.union[1,] %>% str()
og.union[2,] %>% str()
og.union[6,] %>% str()
og.union[8,] %>% str()
```

reformat and combine with rest of the GO table
```{r}
go.table.union <- og.union %>%
  select(-GO_gene_1, -GO_gene_2) %>%
  pivot_longer(c(-og, -GOunion ), values_to = "ID") %>%
  select(-name, -og) %>%
  unnest(GOunion) %>%
  bind_rows(go.table) %>%
  distinct()

gene2GO.union <- go.table.union %>%
  group_by(ID) %>%
  summarize(GOlist = list(GO)) %>%
  mutate(GOlist = set_names(GOlist, nm = ID)) %>%
  pull(GOlist)  
```

```{r}
go.table.union %>% filter(str_detect(GO, "GO:0007140|GO:0007143"))
```


### GO for genes not retained as duplicates in either species 
```{r}
gene.list.neither <- og.category %>% filter(!other) %>% 
  mutate(retained.neither = as.factor(retained.neither*1)) %>% # convert to factor with numeric levels
  pull(retained.neither, name = gene_1) 
str(gene.list.neither)
sum(gene.list.neither=="1")
length(gene.list.neither)
```

```{r}
go.data.neither <- new("topGOdata", ontology = "BP",  description = "Duplicates lost in both", allGenes = gene.list.neither, annot = annFUN.gene2GO, gene2GO = gene2GO.union)
```

```{r}
names(gene.list.neither) %>% duplicated() %>% sum()
```


```{r}
result.fisher.neither <- runTest(go.data.neither, algorithm = "weight01", statistic = "fisher")
result.fisher.neither
```

```{r}
result.table.neither <- GenTable(go.data.neither, result.fisher.neither, topNodes = sum(result.fisher.neither@score < 0.01 )) %>%
  dplyr::rename(P_val=result1) %>%
  mutate(Category = "both lost")

result.table.neither
```

### duplicates retained in both species


```{r}
gene.list.both <- og.category %>% filter(!other) %>% 
  mutate(retained.both = as.factor(retained.both*1)) %>% # convert to factor with numeric levels
  pull(retained.both, name = gene_1) 
str(gene.list.both)
sum(gene.list.both=="1")
length(gene.list.both)
```

```{r}
go.data.both <- new("topGOdata", ontology = "BP",  allGenes = gene.list.both, annot = annFUN.gene2GO, gene2GO = gene2GO.union)
```
```{r}
result.fisher.both <- runTest(go.data.both, algorithm = "weight01", statistic = "fisher")
result.fisher.both
```

```{r}
result.table.both <- GenTable(go.data.both, result.fisher.both, topNodes = sum(result.fisher.both@score < 0.01 )) %>%
  dplyr::rename(P_val=result1) %>%
  mutate(Category = "both retained")

result.table.both
```

### Duplicates retained in S.div


```{r}
gene.list.S_div <- og.category %>% filter(!other) %>% 
  mutate(retained.S_div = as.factor(retained.S_div*1)) %>% # convert to factor with numeric levels
  pull(retained.S_div, name = gene_1) 
str(gene.list.S_div)
sum(gene.list.S_div=="1")
length(gene.list.S_div)
```

```{r}
go.data.S_div <- new("topGOdata", ontology = "BP",  allGenes = gene.list.S_div, annot = annFUN.gene2GO, gene2GO = gene2GO.union)
```

```{r}
result.fisher.S_div <- runTest(go.data.S_div, algorithm = "weight01", statistic = "fisher")
result.fisher.S_div
```

```{r}
result.table.S_div <- GenTable(go.data.S_div, result.fisher.S_div, topNodes = sum(result.fisher.S_div@score < 0.01 )) %>%
  dplyr::rename(P_val=result1) %>%
  mutate(Category = "S_div. retained")

result.table.S_div
```

### Duplicates retained in C_amp.


```{r}
gene.list.C_amp <- og.category %>% filter(!other) %>% 
  mutate(retained.C_amp = as.factor(retained.C_amp*1)) %>% # convert to factor with numeric levels
  pull(retained.C_amp, name = gene_1) 
str(gene.list.C_amp)
sum(gene.list.C_amp=="1")
length(gene.list.C_amp)
```

```{r}
go.data.C_amp <- new("topGOdata", ontology = "BP",  allGenes = gene.list.C_amp, annot = annFUN.gene2GO, gene2GO = gene2GO.union)
```

```{r}
result.fisher.C_amp <- runTest(go.data.C_amp, algorithm = "classic", statistic = "fisher")
result.fisher.C_amp
```

```{r}
result.table.C_amp <- GenTable(go.data.C_amp, result.fisher.C_amp, topNodes = sum(result.fisher.C_amp@score < 0.01 )) %>%
  dplyr::rename(P_val=result1) %>%
  mutate(Category = "C_amp. retained")

result.table.C_amp
```

```{r}
all.results <- bind_rows(result.table.both, result.table.neither, result.table.S_div, result.table.C_amp) %>%
  select(Category, everything())
all.results
```

```{r}
write_csv(all.results, "../output/genespace_no_pinnata/WGD_retention_loss_GO.csv")
```

## KEGG

```{r}
kegg <- read_tsv("../input/Qionghou_Streptanthus_go_kegg/Streptanthus.ko",col_names = c("ID", "KO")) %>%
  mutate(ID = str_c(ID, "-R")) %>% # match format in og.category table
  select(KO, ID) # order needed for enricher
# problems(kegg) # It is OK, just genes without a KO
head(kegg)
```

```{r}
library(KEGGREST)
listDatabases()
term2name <- keggList("ko") %>% as_tibble(rownames = "KO")
head(term2name)
```

```{r}
universe <- og.category %>% filter(other == FALSE) %>% pull(gene_1)
neither <- og.category %>% filter(retained.neither) %>% pull(gene_1)
both <- og.category %>% filter(retained.both) %>% pull(gene_1)
S_div <- og.category %>% filter(retained.S_div) %>% pull(gene_1)
C_amp <- og.category %>% filter(retained.C_amp) %>% pull(gene_2)
```


```{r}
results.kegg.neither <- enricher(neither, universe = universe, TERM2GENE = kegg, TERM2NAME = term2name, pvalueCutoff = .1)
head(results.kegg.neither)
```
 
 

```{r}
results.kegg.both <- enricher(both, universe = universe, TERM2GENE = kegg, TERM2NAME = term2name, pvalueCutoff = .1)
head(results.kegg.both)
```

```{r}
results.kegg.S_div <- enricher(S_div, universe = universe, TERM2GENE = kegg, TERM2NAME = term2name, pvalueCutoff = .1)
head(results.kegg.S_div)
```

```{r}
results.kegg.C_amp <- enricher(C_amp, universe = universe, TERM2GENE = kegg, TERM2NAME = term2name, pvalueCutoff = .1)
head(results.kegg.C_amp)
```
